+++ String.h
@@ -54,9 +54,9 @@ protected :
   //                                Attributes
   // =========================================================================
 	char* tab_;
-	int size_;  //doesn't include de \0
-	int capacity_;  //includes de \0
-	static const int max_size_;	
+	size_t size_;  //doesn't include de \0
+	size_t capacity_;  //includes de \0
+	static const size_t max_size_;	
 
 // ===========================================================================
 //                            Getters' definitions
